Applications.htm & menu.js for 4designbiosciences
This file contains sample code for a HTML website I created for a bioengineering company working on microfluidic devices concerning cancer treatment. The main features in this code are a header menu with drop down capabilities and a picture table with images that can be clicked on to link to other pages in the website. The live page for this HTML document is located at www.4designbiosciences.com/Applications.htm. In addition, I was also the developer for the entire website, all of which is now live at www.4designbiosciences.com.